WL#10923: MTR: Improvements to --bootstrap option
Description:
============
Bootstrap options like innodb_page_size could only be passed to a
test script as a command line option. This meant that a test requiring
a non-default value of a bootstrap option could not be run as a
standalone script and needed the options to be passed on the command line.
The fix for Bug#18184868 provided support for passing bootstrap options
in the opt file.

In this WL, the following bugs were fixed:
1. Bug#25600799: THE METHOD TO PASS BOOTSTRAP PARAMETERS AT COMMAND
   LINE IS CONFUSING
2. Bug#25959942: AVOID REINITIALIZATION OF DATADIR IF PREVIOUS TEST HAS SAME
   BOOTSTRAP OPTIONS
3. Bug#25693748: MTR: PROVIDE A SUPPORT --BOOTSTRAP OPTION IN COMBINATION FILE

The --bootstrap option has been renamed to --initialize.

Change-Id: I5d24c20a725a10147b2d762fc69d9e09d901617e