Bug #29207450 TINYTEXT COLUMN TAKES LONG TIME IN THE GROUP BY OPERATION.
Problem:

The InnoDB intrinsic temp table does not contain an in-place update operation.
An update operation is implemented by doing a delete-mark + insert of a new
record.  Because of this the number of del-marked records in InnoDB temp table
kept on increasing.  Most of the time was spent on scanning the del marked
records.

Solution:

Provide an in-place update operation.

rb#21266 approved by Mayank.