WL#11381: Add asynchronous support into the mysql protocol
C APIs are synchronous, which means client sends a request and waits until
server responds. This makes applications to not do anything until server
sends a response. There can be a requirement where client is not bothered
about when server sends data back, client can submit a query, do some other
tasks and when in need of server response, client can check if server has sent
the response and act accordingly. Thus making a need to implement asynchronous
communication between client and server.

This WL introduces new non blocking version of above C APIs which does not wait
for data to be available on the socket to be read. These APIs will return
immediately with a return status. Return status is used to decide if the API
has completed its operation or needs to be called again at some later point
in time to complete the operation.

RB: 21150
Reviewed-by: Georgi Kodinov
             Ramil Kalimullin