Bug#27394384: RPL TESTS DOES NOT PRESERVE THE INITIAL & FINAL COUNT OF THREAD CONNECTED
Problem:
Many rpl tests have a mismatch in the number of active connections before
and after the test completion.

Analysis:
This is due to the default value of MASTER_HEARTBEAT_PERIOD which is equal
to slave_net_timeout/2 = 60secs. Because of this, even after the slave is
stopped, the master will not close the client session until
MASTER_HEARTBEAT_PERIOD is reached resulting in active Binlog Dump thread
even after the completion of test.

Fix:
Include stop_dump_threads.inc from rpl_end.inc, this will stop the Binlog
Dump thread after the test is finished.