remove my_print_open_files
my_file_opened are now my_stream_opened incremented without locks
so there is a small chance that they will not be statisticly accurate.

As my_end error messages should be accurate, lets not take a chance and
iterate the array, displaying each open file under EXTRA_DEBUG, and
display a warning if a mismatch of files/streams opened occurred.