Bug #29165756 MAKE JOINING GCS MEMBER AUTOMATICALLY DOWNGRADE ITSELF TO THE GROUP'S PROTOCOL
Problem and analysis
========================================================================
WL #11610 "MySQL GCS: Support Data Fragmentation" introduces the notion
of the group's (communication) protocol in GCS.
The group's protocol is specified as a MySQL version, and a group
protocol = X essentially means that the group supports members with
MySQL version >= X.

As part of its joining procedure, GCS validates whether joining servers
are announcing the same protocol as the group's.
If they are, they are allowed in.
If they are not, they are expelled.

WL #11610 introduced a new system variable,
group_replication_communication_protocol_join, that allows users to
specify the protocol announced by a server when it tries to join the
group.
If the protocol announced matches the group's, the server will be
allowed to join.
group_replication_communication_protocol_join's default value is the
MySQL server's version, e.g. 8.0.15.
This means that e.g. downloading MySQL 8.0.15 and trying to add it to an
existing group consisting of MySQL 8.0.14 (or older) members will not
work out of the box.
The user must explicitly set
group_replication_communication_protocol_join to 8.0.14 on the newly
downloaded MySQL 8.0.15 server for it to be accepted in the group.
The user experience can be improved if the MySQL 8.0.15 server adapts
itself to the group's protocol while joining.

Solution
========================================================================
Let the joining member automatically adapt (downgrade) to the group's
protocol.
Automatic adaptation defeats the purpose of the
group_replication_communication_protocol_join system variable, so remove
it.

Implementation pointers
========================================================================
Adapt to group's protocol automatically:
* gcs_xcom_state_exchange.h
* gcs_xcom_state_exchange.cc

Remove group_replication_communication_protocol_join:
* plugin.cc
* gcs_xcom_interface.cc:
* gcs_xcom_utils.cc

RB: 21256
Reviewed-by: Filipe Campos <filipe.campos@oracle.com>
Reviewed-by: Tiago Jorge <tiago.jorge@oracle.com>