Post-push fix for backporting bug#27869124 to Cluster 7.5
The fix for bug#27869124 was back-ported to MySQL Cluster 7.5
to eliminate recurring failures of unit tests, but failed to
compile on some platforms because it uses C++ language features
that are not supported in that environment.

Replace the use of std::mutex in ndbjtie_lib.cpp with NdbMutex.

This change should be null-merged up to mysql-8.0.