Bug#28533993: ERRORS WHEN JOINING ON START MAY LEAD TO A STALE VIEW OF THE GROUP
Issue:
======
The issue is as follows:
1) the joiner joins the group.
2) During view install, it is detected that the server is not compatible
with the group, thence it proactively request to exit the group and it changes
its state to OFFLINE.
3) But the exit group request is not instantaneous. The group must order that
request and then install the view on which the joiner leaves.
4) During that time, between the request and the actual leave, the joiner
has status OFFLINE but still has group membership information, thence the
performance_schema.replication_group_members outcome with local OFFLINE
member and other ONLINE members.

Solution:
=========
We cannot clean the membership information on the joiner after the leave
request, though we can restrict performance_schema.replication_group_members
output to the local member when it is OFFLINE.

RB: 20844
Reviewed-by: Anibal Pinto <anibal.pinto@oracle.com>
             Pedro Gomes <pedro.gomes@oracle.com>