Bug #27383487 TRYING TO ADD A NODE BACK AFTER A MAJORITY LOSS MAY LEAD TO INCONSISTENCIES
Problem and analysis
========================================================================
After analysing the source code and discussing on Slack, we concluded
that a node may rejoin the group "out of thin air" after it crashes,
i.e. become a productive member of the group without going through the
remove+add node path after it crashes.
This is due to the way XCom members get added to the group:

1. One sends an add_node to the existing XCom group with the address of
   the joining member.
2. The group agrees to add the joining member to the group, it is added
   to the membership.
3. The existing members begin sending messages to the joining member.
4. Upon receiving messages "from out of nowhere," the joining member
   replies with "I'm new, please send me stuff so I can be up to date."

If a node crashes and comes back before TCP connections are deemed over,
it is possible that it begins receiving messages intended for its
previous, pre-crash incarnation. As per step 4, the post-crash
incarnation will request to be updated and begin processing messages
successfully. Since the post-crash incarnation has forgotten about
previously accepted values, it may become part of a quorum that leads
to the acceptance of a different value. Boom!

Since the pre-crash incarnation may have accepted a value for a
given synod but the post-crash incarnation has forgotten that fact, the
post-crash incarnation will fail to propagate the previously accepted
value to a higher ballot. Since majorities can overlap on a single node,
if the overlap node is the post-crash incarnation which has forgotten
about the previously accepted value, a higher ballot proposer may get
a different value accepted, leading to conflicting values to be accepted
for different proposers, which is a violation of the safety requirements
of the Paxos protocol.

Solution
========================================================================
The overall is idea is to prevent post-crash incarnations from being
productive Acceptors in Paxos.

The concrete patch is:

* Augment the "I'm new, please send me stuff so I can be up to date"
  message (need_boot_op) with the joining member's incarnation
  identifier.
* Let the recipient of the need_boot_op message check whether the
  incarnation identifier in the message matches the incarnation
  identifier in the membership information:
  a. If the incarnation identifiers match, the recipient of the
     need_boot_op message processes it.
     (It is indeed a valid request from a fresh member that was just
     added via an add_node command.)
  b. If the incarnation identifiers *do not* match, the recipient of the
     need_boot_op message ignores it.
     (It is a post-crash member attempting to rejoin the group out of
     thin air.)
* Prevent non-booted members from participating in Paxos as Acceptors,
  i.e. let them ignore prepare_op and accept_op messages.

Reviewed-by: Ole-Hjalmar Kristensen <ole-hjalmar.kristensen@oracle.com>
Reviewed-by: Tiago Jorge <tiago.jorge@oracle.com>
RB: 21217