Bug #29039732 	SHOW BINLOG EVENTS FROM <BAD OFFSET> IS NOT DIAGNOSED - post fix
Problem
=======
There is an assert failure on verifing that the starting offset (232)
of an event after the invalid position 156 is displayed.

Analyze
=======
The length of a Gtid log event has a different length in different
platforms in binary log file. So the end position of the Gtid log
event is 232 in some platforms and 234 in others. Which causes the
above assertion failure.

Fix
===
Remove the assertion on the event after the Gtid log event binary
log file, and add an assertion on a event before it.

RB: 21293