Bug#28825718 - ASSERTION FAILURE: TRX0REC.CC:NNN:N_IDX > 0 WHILE DOING REPLACE/INSERT
If a try to create a single index on a virtual column fails, then after
the rollback, the column actually does not belong to any index and its
ord_part should be 0. However, since the rollback was affected by some
concurrent access to the table, dropping the half created index has to be
delahyed until next DDL. The problem is that on next DDL, it just drops
the half-done virtual index, but not updates the ord_part of the column,
which leas to a mismatch between the ord_part of the column and the virtual
index list of this column. Thus finally the assertion is triggered.

The solution is that the ord_part of all affected columns should be reset
after the rollback of the half created virtual indexes, as what it is done
for normal DROP INDEX. In this way, the ord_part and the virtual index
list of the same virtual column are consistent.

RB: 21257
Reviewed-by: Kevin Lewis <kevin.lewis@oracle.com>