Bug#29196401: PB2 FAILURE OF NDB_DISCOVER_DB + NDB_FK_LOCASE
Issue:
-----
With the changes made in WL#10923, the master is reinitialized
if the slave is needs reinitialization and vice versa. NDB tests
which start two mysqld instances do not have the concept of master
and slave, and hence the second server is not reinitialized, even
if the first one is.

Fix:
---
In the loop which checks if any of the servers need reinitialization,
there was an exit if even one server needed reinitialization, and
so the second server's need for reinitialization value was undefined.
Now, for NDB tests, the loop continues to execute and returns
the correct value if reinitialization is required.

Change-Id: I704c354ea2575f4c59283cee38d38003ec03bfcc