Bug#28772867 - RECENTLY ADDED NDBMTD HANGS IN PHASE 4 DURING STOP
Problem: When newly added nodes are started almost in parallel, i.e.,
the start-phases are interleved, one of the newly added node looses
information on the other newly added node, which is going to be in its
node group. This lacking knowledge about the handover node causes
stopping the newly added node to hang.

Parallel starts cause the execution order of two signals -
INCL_NODEREQ and READ_NODESCONF - to switch. Thus the last signal
execution overwrites the handover node info.

This problem does not occur when the newly added nodes are started one
after another.

Fix: The execution of READ_NODESCONF should merge the started nodes
info with c_alive_nodes instead of overwriting it.

Reviewed-by: Mikael Ronström <mikael.ronstrom@oracle.com>