Bug#29221385: INNODB GOT ASSERT FAILURE WHILE FIGURING OUT SPACE ID AT STARTUP
This bug is reported on a file-per-table tablespaces that uses compression
and is apparently corrupted such that the function get_tablespace_id()
does not succeed in finding the same space_id on the first four pages,
so it proceded to the more involved approach of trying multiple page sizes.
When this happens, it is not valid to assert that each page read as a
compressed page will have a valid version number in its header.

So that assert is deleted in this patch.

Approved in RB#21310 by Rahul Agarkar