BUG#29242900: MTR: SANITIZER REPORT IS REPEATED IN MTR'S OUTPUT LOG
Description:
============
When --sanitizer option is enabled, sanitizer report generated is
repeated in MTR's output log.

Issue:
======
MTR maintains an array of server error log files to search for any
sanitizer related errors/warnings. However there can be duplicate
entries in this array causing the same error log file to be parsed
multiple times and generating the repeated reports in the log.

Fix:
====
Maintain a hash of error log files instead of an array. The error log
file path will be the key values in the hash which will take care of
avoiding the duplicate entries.

Change-Id: I30f9fb5d910c1158e88653e4e415819dfc86238a