Bug#29240455 NDBINFO_UPGRADE TEST FAILURE IN PB2/WINDOWS
The test case when performing upgrade, redirects the output to a log file
and cleans it at the end of the test case using --remove_file command.
This command acts in a different way in windows. On platforms other than
Windows, the command gets executed and deletes the log file instantly. In
Windows the command returns success, but can not delete the file until the
mysqld is stopped. But, before mysqld is stopped, MTR calls Check-test case,
which finds the log file in the tmp directory and fails the test case.

As the standalone mysql_upgrade tool is merged to mysqld, usage of log-file
is now not required and hence removed.

Change-Id: I372df9aaabf4fc16d7d62fe01616b331f95e673f