Bug#28932965 WRONG NDB_RESTORE BINARY CALLED IN TEST
Fixes PATH search order that was causing wrong ndb_restore binary to be picked
up when running "atrt-testBackup -n Bug19202654" when test was run in supra07
machines which have a system installed version of cluster.

Several failures were reported while running atrt-testBackup -n Bug19202654
T1 test due to unknown option "--error-insert" for ndb_restore, even though
this option is available in trunk. Initial investigation showed that a
different binary is being used which is manually installed on the
machine. But, still PATH environment variable should have the
appropriate path to the binary before the system path. This was in missing
atrt-testBackup where path for the binary was added after the system
path which is "/usr/bin" in this case. So, $MYSQL_BASE_DIR/bin is added before
$PATH in order to make the test look at the proper path before looking
for system installed version.

Approved by: Tiago Alves <tiago.alves@oracle.com>