BUG#28642608: UPGRADE FROM 5.7 -> 8.0 DOES NOT DETECT INCONSISTENT DD
Analysis
========

Upgrading from MySQL 5.7 to MySQL 8.0 results in an inconsistent data
dictionary when the tablespace files created per table(InnoDB SE) or for
the shared tablespaces are deleted after 5.7 server is shutdown.

In MySQL 5.7, 'innodb_file_per_table' is enabled by default, hence a
tablespace file is created per table. The entry for the tablespace is
maintained in the SYS_TABLESPACES. During the upgrade the entries are
migrated from SYS_TABLESPACES to DD:TABLESPACES. The entry for the
tablespace remained in SYS_TABLESPACES since the file was deleted
after the server was shutdown. Thus the entry was migrated from
SYS_TABLESPACES to DD:TABLESPACES creating an inconsistent data
dictionary after upgrade.

Fix
===
During the upgrade of tablespace files, check for the existence of
the tablespace file before updating the DD:TABLESPACES. Also the
debug assert has been removed to ensure appropriate error is reported
when the general tablespace file is removed.

ReviewBoard: 21274