Bug #28893633 DBTUP "CHECK FALSE FAILED" CRASHES ON SEVERAL DATA NODES
Problem

A race between ACC and LQH when different operations in a transaction
on the same row are concurrently being PREPARED and ABORTED.
This race can result in TUP attempting to PREPARE an operation when
a preceding operation has been aborted, which is unexpected, and can
result in undefined results.

Solution

Modify ACC and LQH to check that all dependencies are still valid
before proceeding to PREPARE an operation.
Make the design assumption that abort of an operation will be
followed by abort of all its dependencies more concrete.
This fix supersedes the previous fix in this area pushed as part
of bug#28500861.

Test coverage is provided by new testcase
testBasic.cpp -n AbortRace

Approved by: Ole John Aske <ole.john.aske@oracle.com>