Bug #28995220  SEGFAULT DURING SERVER START IF BINLOG INDEX VALUES ARE EXHAUSTED
Problem
=======
Restart server with exhausted binlog index value causes a segmentation fault.

Analyze
=======
The server tries to report an error on the session by a NULL THD when
restarting it with exhausted binlog index value. The NULL THD causes
the segmentation fault.

Fix
===
Don't report an error on the session by a NULL THD. The server would
abort with errors regardless of the value of binlog_error_action
while restarting it with exhausted binlog index value.

RB: 21299