Bug#28986737: RENAMING AND REPLACING MYSQL.USER TABLE CAN               LEAD TO A SERVER CRASH
Description:- Server crashes when mysql user table is
damaged with some missing columns.

Fix:- A check is introduced for confirming the required
columns are present. An error is printed in the server log
if mysql user table is damaged.

NOTE:- This issue is not repeatable in 5.7+ since a similar
check is already done inside "acl_load()" for 5.7.

RB#21154