Bug #29270297: ADD 8.0.16 TO LIST OF SUPPORTED SERVER VERSIONS
During upgrade, there is a check if upgrade is allowed from the server
version of the data directory. To do this, there was a list of supported
versions. Since this will mean adding a server version during every
release of a server, this patch inverts the logic and checks for upgrade
with the help of a list of server versions from which upgrade is not
supported.

Change-Id: I7b020ab7044137952c69e9e3633cfba7969fea9d