Bug#29262359 - MTR TEST NDB_DD_BASIC FAILS WITH " TABLE IS FULL"
Problem: When inserting rows into a disk table with blob columns,
'Table is full' error is returned.

Reason: In the test, each ndb node has 4 ldms and the table has 1
fragment per ldm. That means, the table needs minimum 8 extents (4
backup fragments) on one data node. The test created a tablespace with
16 extents. Note: 4% of a table space is reserved for restart and
there are some overhead to store data file meta data.

The test creates and drops tables before the failure. An LCP started
before a couple of drop-tables has not completed when populating the
newly created table failed with 'Table is full'. This is because, the
extents used by the dropped tables were not freed.

The timing of LCP completion explains the intermittent nature of
the failure.

Fix: Increase the tablespace size of the test.

Reviewed-by: Ole John Aske <ole.john.aske@oracle.com>