Bug#28940361 QUERY AGAINST FULL TEXT INDEX WITH ORDER BY SILENTLY FAILS
Problem:
The FTS query with an ORDER BY clause silently returned because of searching in
an incorrect index. The query returns silently if it meets the following
criteria:
1) The table should should have an implicit FTS index
2) The table should be created with a non-default collation
3) The query should have an ORDER BY clause
4) The sort buffer size should be inadequate to hold the sorted keys

The root cause of the problem is the incorrect index set while
looking for the rows from the clustered index. The primary index
needs to be active while searching the rows using the row ids. However,
in this case, the index used is the FTS_DOC_ID_INDEX.

This problem does not happen if the sort buffer is large enough to hold
all the sort keys. This is because the index is reset to the primary
index while initializing the iterator to drain the sort keys from the
sort buffer and, hence the active index is correctly set to primary
index while retrieving rows from the primary index.

Solution:
Reset the active index to the primary index while initializing the
file sort iterator.

Fix:
Reset the index to the clustered index in
SortFileIndirectIterator::init() method.

RB: 21334

Reviewed By: Steinar Gunderson (steinar.gunderson@oracle.com)

Change-Id: I18ba20c3f9c058e27f5767ddf30ec77e1197aec0