Bug#29272683 CLEANUP NULLABILITY AND ER_WARN_ALLOWED_PACKET_OVERFLOWED FOR STRING FUNCTIONS
In non-strict SQL mode, certain string functions will return NULL to
indicate that the result is too large for the result buffer. This may
result in inconsistent behaviour (e.g. incorrectly sorted output) when
'maybe_null' is set to 'false'.

The fix is to set 'maybe_null' to 'true' if the result string may be
longer than max_allowed_packet.

Change-Id: I8236cd558213a2b845545027ad9d1a427cf6b03f