Bug #29278753: EXPLAIN FORMAT=TREE MISSES SOME SUBQUERIES IN PROJECTIONS
EXPLAIN FORMAT=TREE does not always show all subqueries in projections. In
particular, those that are materialized during execution (due to ordering or
the likes) are not displayed. Fix it by looking through the QEP_TABs
corresponding to temporary tables for such items.

The format is still marked as experimental and subject to change, so it's not
a critical issue, but it's very useful during debugging to see all the plans.

Change-Id: I95c345a6c39ebbbadac2e5680e2c5c27eda780d1