Bug#29231709 NULL VALUES (still) NOT FILTERED OUT BY PUSHED CONDITIONS
Follow up bug to bug#28643463:  NULL VALUES NOT FILTERED OUT BY PUSHED CONDITIONS

Original patch did not take into account that the interpreter code for
a predicate of the form '<value> op <field>' was always generated with the 'field'
as the first operand. Thus the 'op' was swapped to the 'inverted op',
such that the generated code would be '<field> swap(op) <const>'.
The logic for adding a '<field> IS NULL' check or not, did not
take the possible 'swap(op)' into consideration.

This patch add the function Ndb_item::swap() which is now
used when deciding whether to add 'IS NULL' checks or not.

Reviewed by : Martin Skold <Martin.Skold@oracle.com>