Bug#29264755 EMPTY CLUSTER-METADATA-SERVERS IN STATE.JSON Bug#29271620 CANNOT RE-BOOSTRAP MYSQL ROUTER WHEN IT IS RUNNING SINCE 8.0.14
When the router while refreshing its metadata could not
connect to any of the metadata servers (because of some
network issues etc.) it would clear the metadata servers
list in its dynamic state file. That would result in it
not being able to access cluster metadata when restarted.

Another side effect was Bug#29271620: While doing
the second bootstrap using the same output directory,
the router removed the old account in the cluster so
the running router instance started receiving 'Access
denied' while trying to query the metadata. Because of
BUG#29264755, the running instance overwrites the dynamic
configuration file state.json with empty metadata-servers list.

This patch prevents the router from storing empty metadata-server
list in the state file.

Reviewed-by: Pawel Mroszczyk <pawel.mroszczyk@oracle.com>
RB: 21381