Bug#29283147 - MTR: NDB_BACKUP_ALTER: 'ALTER TABLE..' SUCCEEDED - SHOULD HAVE FAILED WITH 1296
The test intend to start a backup (BU), pause it, then perform the
alter, unpause the BU to continue. The alter should fail since the
BU holds the 'lock' (m_read_locked) in dict.

Problem: The BU has not been started when the alter is performed
due to timing. Thus the alter was started before BU and succeeded.

Fix: Make the 'start backup' to return only after the backup is
started, instead of returning immediately.

Reviewed-by : Frazer Clement <frazer.clement@oracle.com>