BUG#29270712 PERFORMANCE SCHEMA RWLOCK INSTRUMENTATION LACKS UNLOCK DETAILS
Before this fix, the performance schema instrumentation for RWLOCK would:
a) name priority read write locks as "rwlock"
b) not collect information about the kind of unlock operation performed,
   in unlock_rwlock()

a) is misleading, because there are 3 kinds of read-write locks supported:
- plain read write locks, named "rwlock",
- priority read write locks, also named "rwlock", causing confusion,
- shared exclusive read write locks, named "sxlock".

With this fix, priority locks are named "prlock",
as in "wait/synch/prlock/sql/MDL_lock::rwlock".

b) is a semantic problem that affects recursive locks such as shared
exclusive locks.
When a shared exclusive lock is locked multiple times,
and when an unlock operation is instrumented,
the instrumentation does not know what lock is released,
and the resulting lock state is ambiguous.

The fix is to provide an extra parameter in the unlock_rwlock()
instrumentation, to clarify what actual operation is performed.

As a result, the instrumentation interface for unlock_rwlock() changes,
so a new version of the instrumentation service (psi_rwlock_v2) is now
exposed.
The previous instrumentation, psi_rwlock_v1, is now obsolete.

Approved by: Chris Powers <chris.powers@oracle.com>