WL#12865 SUPPORTS ATRT SKIPPED TESTS
Adds support to "NDBT_ProgramExit: 4 - Skipped" test results used to
skip test execution.

Our current upgrade/downgrade test suite is running tests for
non-supported scenarios which report failure. To handle this situation
changes were done in NDBT library and testUpgrade to be able to skip a
test execution.

This patch adds support in ATRT to recognize a skipped test result.

Changes in atrt-analyze-result.sh were done to consider skipped test
results in addition to ok and failed results. The overall test result is
computed by following the priority Failures -> Skipped -> Ok. I.e, if
any failure is detected the test fails, if no failures are detected but
any skipped step is, then the test is skipped, finally if there's no
failures and no skipped steps and there's at least one OK then the test
succeeds. NOTE that if there's no Failures/Skipped/Ok, this means that
test didn't output a valid result and the test should be considered as
Failure.

Changes in atrt were done to recognize a skipped test result and print
this new status in the log and updating code such that is not treated
like a test failure.
Additionally updated ErrorCodes enumeration to contain all status
tracked during test execution. This enumeration is used both for atrt
process execution status and test status. Further refactoring should be
considered to separate this enumeration in two. Furthermore, could be
significantly cleaned but this will be left to be done in the context of
subsequent tasks that will change ATRT main algorithm.

Change-Id: I9e8c7dd50604c0d7cbd4fe35166af5f1f938716e