BUG#29140298 - `OPT_EVENT_SCHEDULER == EVENTS::EVENTS_ON ||                OPT_EVENT_SCHEDULER == EVENTS::EVEN
When MySQL server is started with --event-scheduler=DISABLED
debug assert which checks event-scheduler state ON or OFF fails
during event subsystem initialization.
The root cause here is, the events should not be initialized
when server is started with the events DISABLED. But currently
condition to check for events DISABLED is missing during event
subsystem initialization and proceeding with the initialization
hits the assert on debug builds.

The patch fixes the issue by checking for Events::EVENTS_DISABLED
and not proceeding with the events initialization.

Reviewed-by: Praveenkumar Hulakund <praveenkumar.hulakund@oracle.com>