BUG#29304076 OPTIMIZE ROW UNPACKING BY NOT COMPUTING FUNCTIONAL INDEXES FOR THE BEFORE-IMAGE
Description
-----------

In RBR, when unpacking a row for an `Update` or `Delete` event, the applier
deserializes the before-image and the after-image. During this process, for
both images, if the target table has functional indexes, the applier tries to
compute the index value. This is unnecessary for the before-image since those
values aren't used.

Analysis
--------

Row unpacking could be optimized by not calculating values for functional
indexes when upacking the before-image parte of the row.

Fix
---

Test if it is the before-image that is being unpacked and if so, don't compute
functional indexes values.

ReviewBoard: 21416
Reviewed-by: Sven Sandberg <sven.sandberg@oracle.com>