Bug#29276074 CRASH IN ITEM::DELETE_SELF()
The windowing bit aggregates use a dynamically allocated counter
array. This is allocated based on Item::max_length. In the repro,
max_length is wrong, in that it does not allow for the extra escape
characters that may be needed. In the repro, an escape character is
added.

So, when Item_sum_bit::add_bits tried to access the counter array
using the actual string, it was longer than max_length, and the access
was out of bounds for the array, leading to the crash. Since writing
happens beyond allocated memory, even if no crash is seen, running with
ASAN should reveal the issue.

Solution: fix max_length for QUOTE. Added ASSERT that the actual
length of the string is within bounds of the counter array.

The determination of max_length for QUOTE (Item_func_quote) was
broken, fixed in Item_func_quote::resolve_type.

In addition to the repro, added a test case with CREATE TABLE AS
SELECT QUOTE(), which used to give an error due to this wrong
max_length.  This now works correctly. Metadata for QUOTE will now be
more pessimistic, which impacts some earler tests' canons.

Change-Id: Ica52ff612676218c659130a721f5faccad24d6f7