Bug#28817441: CRASH IN FIND_ACL_USER()
Description:- Server crashes when trying to drop a
user which has granted a particular role.

Analysis:- When trying to drop a user having a role,
server crashes inside  'find_acl_user()' when it tries
to match the hostname of the role.

Fix:- A check is introduced to resolve this issue.

RB#21223