Bug#29283055: CREATE TABLE IF NOT EXISTS .. SHOULD NOT ALWAYS FAIL ON SQL_REQUIRE_PRIMARY_KEY
Problem: CREATE TABLE IF NOT EXISTS would fail even if the table
already exists, if the new table definition does not have a PK and the
system variable sql_require_primary_key is set.

Solution: Delay the call to mysql_prepare_create_table(), which
performs this and other checks, until the check for an existing table
has been done. This makes sense as the work done in
mysql_prepare_create_table() is wasted when the table already exists,
since create_table_impl() then returns early, either with error or
success.

Change-Id: I6e8bf6b0bf6f41426e7d391db7a9bf8d8611d7b3