BUG#29292860: ERROR 197 FROM SE WHILE MIGRATING TABLESPACES               IN AN UPGRADE
Analysis
========
Upgrade from MySQL 5.7 to 8.0 on windows platform fails after
the patch for:
BUG#28642608: UPGRADE FROM 5.7 -> 8.0 DOES NOT DETECT INCONSISTENT DD
While checking for tablespace file's existence, the file is opened in
a read_only mode to check for its existence. On windows, attempting to
open the file when it is already accessed caused access share violation,
preventing upgrade by reporting an error.

Fix
===
Check if the tablespace is already accessed during upgrade before attempting
to open the file in read_only mode to check for its existence.

ReviewBoard: 21429