Bug#29306411 ASAN BUILD FAILS IN PB2 AFTER COMPILER UPGRADED TO CLANG-7
Aligned_alloc(alignment, size) are used for allocating memory for block classes
derived from SimulatedBlock. Aligned_alloc() expects the requested size to
be an integral multiple of the alignment(64 bit). But size of most of the
inherited class members were not 64 bit aligned. So when an instance is created,
aligned_alloc() is called with a size, which is not multiple of 64 and gives error.

Fixed the issue by making Class SimulatedBlock to be 64 bit aligned and thereby
making all the classes inherited from it also to be 64 aligned.

Change-Id: I04cc76f99ed3ed556f0ce5a4bc60c87574a5f021