Bug#27349236 - NODE MAY RECEIVE A LOCAL VIEW AFTER BEING KICKED OUT OF THE GROUP
Description
-----------
A node may still deliver a local view after being expelled from a group.
If this happens, the error log may show a message stating that the member
has been expelled from the group followed by another message stating that
the node has resumed regular operation. While the first message is correct,
the second is not. This is possible because local views are not totally
ordered and can be individually triggered by the node after it receives a
(totally ordered) global view in which it has been expelled.

Solution
--------
Check that node has not been expelled before delivering a view to the
event listeners.

RB: 21300
Reviewed by: Filipe Campos <filipe.campos@oracle.com>
Reviewed by: Tiago Jorge <tiago.jorge@oracle.com>