BUG#29323596 [3/3] ROUTER FAILS TO REPORT ERRORS WHEN RUNNING AS A WINDOWS SERVICE
When Router is started, it logs its messages to console until a moment
when it switches to sink(s) defined in the configuration file. However,
when running as a Windows service, the console doesn't exist, thus
anything logged there will be lost and never seen by the user.

This patch fixes this issue when running as a Windows service by logging
to both console and Windows Eventlog until we make the switch to the
sink(s) defined in the configuration file.

RB#21440