BUG#29053128: SEGFAULT IN GROUP_PARTITION_HANDLING.CC POST STOP G.R.
Description:
When issuing a STOP GROUP_REPLICATION while the member is trying to
assess whether or not it lost majority (i.e. during the execution part
of the Group_partition_handling thread loop), the server encountered a
segfault on access to the Group_partition_handling mutex.
This happens because the Group_partition_handling instance was
destroyed (and in the dtor it destroyed the mutex) before stopping the
thread.

Solution:
The issue was to terminate the group partition thread in
terminate_plugin_modules() before destroying the instance.

RB: 21363