Bug #29324132 	WL#9508 INTRODUCED NON-IDIOMATIC POTENTIALLY-BROKEN C MACROS
As the bug report mentions, there are multiple ways to fix this macro.

The one I prefer is to change these macros into inline functions.
This solution enforces its use with consistent types.  But it also
exposes the fact that fsp flags are sometimes declared as uint32_t
and sometimes ulint.  In older versions they were always ulint as
well as the table flags and related 32 bit numbers. The type ulint
has an inconsistent number of bytes.  The process on converting all
table flags to uint32_t is incomplete until now.  This patch finishes
that conversion.

Approved by Rahul Agarker in RB#21445