Bug#29317684 REPLICATION IS SENSITIVE TO ORDER OF HIDDEN COLUMNS FOR FUNCTIONAL INDEXES
When creating a functional index part, the server creates a hidden
virtual generated column which is in turn indexed. The new column is
placed after the last column in the table, and regular columns that
are added later will be placed after the hidden column again. This
has several problems:

1) It creates "holes" in the column position index. When selecting
from INFORMATION_SCHEMA.COLUMNS, you might see column positions
1, 2, 4 and 5, where 3 is the hidden generated column.

2) If you dump a table where you have a hidden column in the middle
and apply the CREATE TABLE statement, the hidden columns will now be
located at the end. This is especially problematic for replication
where you could end up in a scenario where the columns have a different
location on the master and slave. Note that this can be weird in a
non-replication setup as well where the user would see column indexes
1, 2 and 4 before dump/restore, and 1, 2 and 3 after a dump restore.

This patch forces all hidden generated columns to always be located
at the end, sorted by their column name, so that the order of the
hidden generated columns are deterministic regardless of which order
they were added in. Note that the column names for the hidden columns
are deterministic, so this will ensure that;

a) Master and slave will always have the same column order for both
   visible and hidden columns
b) We will never get these "holes" in the column position index as
   described above.

The effect of this patch is that if the user creates a new regular
column, it is always added before the first hidden column. And if a
new functional index part is added, the new hidden column is
inserted in the correct position. Note that there is a preparatory
patch for this bugfix that allows this re-ordering to be executed
both INPLACE and INSTANT.

There is one exception to this behavior; if we have a replication
setup where the master is a MySQL version where functional index
columns aren't sorted last, the slave will not do it either. If the
slave did it, we would get an mistmatch between the column order
on the master and the slave, and the bug would stil be present.

An notable behavior is that if a user bootstraps a MySQL server
(running the latest version) from a dump file, the slave would get
a table structure where hidden columns are sorted correctly.
If said slave then connects to a master that does not sort hidden
column correctly, the replication is broken. So a working replication
setup heavily relies on the fact that table on the slave is produced
by replicating DDL from old master or as result of binary upgrade
from old server version.

When upgrading from old version to a new version, there are two
scenarios involving hidden columns an user can encounter:

1) Consider a table that has one regular column as the first column,
   and two hidden columns that are in the wrong order. The first
   ALTER TABLE on this table will re-order the two hidden columns,
   and it might be executed both INPLACE and INSTANT, depending on
   what the user is altering.

2) Consider a table that has one regular column as the first column,
   a hidden column as the second column, and a regular column as the
   third column. The first ALTER TABLE on this table will re-order
   column two and three, so that the hidden column is placed at the
   end of the table. This might also be executed both INPLACE and
   INSTANT, depending on what the user is altering.

So for such cases, the first ALTER TABLE on the table will implicitly
reorder hidden columns. This might (but not necessarily will) change the
ALGORITHM required for executing the ALTER TABLE (for instance,
the combination of DROP COLUMN and reordering hidden columns requires COPY).

Note that the table structure will not be changed automatically during
upgrade; that must be performed by the user explicitly by executing
an ALTER TABLE statement.

Change-Id: Ic376bb1fcdd4ba8aeb16292b4489f19c5772550e