Bug#29327469: MTR SHOULD THROW AN ERROR IF UNINSTALL PLUGIN FAILS               FOR SECONDARY ENGINE
Issue:
------
If the secondary engine plugin cannot be uninstalled successfully
between test runs, then the next test will fail because the plugin
could not be installed. If the uninstall error is caught in the
current test, then it will not affect subsequent tests.

Fix:
----
If the secondary engine plugin could not be uninstalled,
because it is busy, or for any other reason, the current test will
fail.

Change-Id: I89ca753efb0e3390d3a1a3d9c9a5766090447faa